requirements.txt
flask
pandas
werkzeug
gunicorn
psycopg2-binary
SQLAlchemy>=2.0
psycopg2-binary>=2.9
Pillow
requests